* made encoding as standard * namespace improvements
