+++ src/di/inject.rs
@@ -1,5 +1,4 @@
-﻿use crate::di::Container;
-
+﻿use super::Container;
 use std::{io::Error, future::Future};
 use futures_util::future::ok;
