+++ src/middleware.rs
@@ -6,9 +6,9 @@ use std::{future::Future, sync::Arc};
 pub use http_context::HttpContext;
 
 use crate::{
-    status,
     App,
     HttpResult,
+    status
 };
 
 #[cfg(any(
